Feature/reservation price

+++ src/main/java/com/SleepUp/SU/utils/email/EmailService.java
@@ -4,22 +4,17 @@ import com.SleepUp.SU.accommodation.entity.Accommodation;
 import com.SleepUp.SU.reservation.entity.Reservation;
 import com.SleepUp.SU.user.entity.User;
 import jakarta.mail.MessagingException;
-import jakarta.mail.internet.MimeMessage;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Service;
-import org.thymeleaf.context.Context;
-import org.thymeleaf.spring6.SpringTemplateEngine;
+
+import java.math.BigDecimal;
 
 @Service
 public interface EmailService {
     void sendWelcomeEmail(String toEmail, String userName) throws MessagingException;
 
-    void sendOwnerReservedNotification(User guest, Accommodation accommodation, Reservation reservation, double amount) throws MessagingException;
+    void sendOwnerReservedNotification(User guest, Accommodation accommodation, Reservation reservation, BigDecimal amount) throws MessagingException;
 
-    void sendReservationConfirmationEmail(String toEmail, String userName, String accommodationName, String location, String checkInDate, String checkOutDate, double amount) throws MessagingException;
+    void sendReservationConfirmationEmail(String toEmail, String userName, String accommodationName, String location, String checkInDate, String checkOutDate, BigDecimal amount) throws MessagingException;
 
     void sendReservationReminderEmail(String toEmail, String guestName, String accommodationName, String location, String checkInDate, String checkOutDate) throws MessagingException;
 
@@ -31,7 +26,7 @@ public interface EmailService {
 
     void sendCancellationNotificationToOwnerEmail(String toEmail, String userName, String accommodationName, String location, String checkInDate, String checkOutDate) throws MessagingException;
 
-    void sendReservationConfirmationEmail(User guest, Accommodation accommodation, Reservation reservation, double amount) throws MessagingException;
+    void sendReservationConfirmationEmail(User guest, Accommodation accommodation, Reservation reservation, BigDecimal amount) throws MessagingException;
 
     void sendReservationReminderEmail(User guest, Accommodation accommodation, Reservation reservation) throws MessagingException;
