+++ src/test/java/com/SleepUp/SU/utils/EmailServiceImplTest.java
@@ -17,6 +17,7 @@ import org.springframework.mail.javamail.JavaMailSender;
 import org.thymeleaf.context.Context;
 import org.thymeleaf.spring6.SpringTemplateEngine;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import static org.mockito.ArgumentMatchers.*;
@@ -81,7 +82,7 @@ class EmailServiceImplTest {
         when(templateEngine.process(eq("NotificationReservationOwner"), any(Context.class)))
                 .thenReturn("<html>Owner Notification</html>");
 
-        emailService.sendOwnerReservedNotification(guest, accommodation, reservation, 3);
+        emailService.sendOwnerReservedNotification(guest, accommodation, reservation, BigDecimal.valueOf(3));
 
         verify(mailSender).createMimeMessage();
         verify(mailSender).send(mimeMessage);
@@ -101,7 +102,7 @@ class EmailServiceImplTest {
                 "Madrid",
                 "2025-10-01",
                 "2025-10-05",
-                3
+                BigDecimal.valueOf(3)
         );
 
         verify(mailSender).send(mimeMessage);
@@ -114,7 +115,7 @@ class EmailServiceImplTest {
         when(templateEngine.process(eq("ConfirmationGuest"), any(Context.class)))
                 .thenReturn("<html>Confirmation</html>");
 
-        emailService.sendReservationConfirmationEmail(guest, accommodation, reservation,3);
+        emailService.sendReservationConfirmationEmail(guest, accommodation, reservation,BigDecimal.valueOf(3));
 
         verify(mailSender).send(mimeMessage);
         verify(templateEngine).process(eq("ConfirmationGuest"), any(Context.class));
@@ -198,7 +199,7 @@ class EmailServiceImplTest {
                 .thenReturn("<html>Email</html>");
 
         emailService.sendWelcomeEmail("test@test.com", "Test");
-        emailService.sendReservationConfirmationEmail(guest, accommodation, reservation,3);
+        emailService.sendReservationConfirmationEmail(guest, accommodation, reservation,BigDecimal.valueOf(3));
         emailService.sendCancellationConfirmationEmail(guest, accommodation, reservation);
 
         verify(mailSender, times(3)).send(mimeMessage);
